src/requirements.txt
# =============================================================================
# DEPENDENCIAS DE SYNAPSE AI
# =============================================================================

# Framework web
Flask==3.0.3
Werkzeug==3.0.3

# CORS para API
flask-cors==4.0.1

# Rate limiting para protección contra abuso
Flask-Limiter==3.5.0

# Headers de seguridad HTTP
flask-talisman==1.1.0

# Cliente de OpenAI
openai==1.51.2

# Gestión de variables de entorno
python-dotenv==1.0.1

# File locking para concurrencia
filelock==3.16.1

# Validación y serialización de datos
pydantic==2.9.2
pydantic-settings==2.6.0

# Dependencias indirectas (evitar conflictos)
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
requests==2.32.3
urllib3==2.2.3

# =============================================================================
# DEPENDENCIAS DE TESTING (OPCIONAL)
# =============================================================================
# Descomentar para instalar herramientas de testing:
# pytest==7.4.3
# pytest-flask==1.3.0
# pytest-cov==4.1.0
